refactor: update color of message

+++ app/src/main/java/com/example/product_sale_app/ui/chat/ChatListActivity.java
@@ -1,4 +1,4 @@
-// com/example/product_sale_app/ui/chat/ChatListActivity.java
+// src/main/java/com/example/product_sale_app/ui/chat/ChatListActivity.java
 package com.example.product_sale_app.ui.chat;
 
 import android.content.Intent;
@@ -30,37 +30,41 @@ public class ChatListActivity extends AppCompatActivity {
         rvChats = findViewById(R.id.rvChats);
         rvChats.setLayoutManager(new LinearLayoutManager(this));
 
+        // Pass initial empty list + listener
         adapter = new ChatListAdapter(new ArrayList<>(), boxId -> {
-            // onClick: open ChatActivity for this box
             Intent i = new Intent(this, ChatActivity.class);
-            i.putExtra("boxId", (int)boxId);
+            // NO CAST needed — boxId is already an int
+            i.putExtra("boxId", boxId);
             startActivity(i);
         });
         rvChats.setAdapter(adapter);
 
-        // Retrofit + repo
+        // Build Retrofit + repo
         HttpLoggingInterceptor log = new HttpLoggingInterceptor()
                 .setLevel(HttpLoggingInterceptor.Level.BODY);
         OkHttpClient client = new OkHttpClient.Builder()
-                .addInterceptor(log).build();
-
+                .addInterceptor(log)
+                .build();
         ChatApiService api = new Retrofit.Builder()
                 .baseUrl("http://10.0.2.2:5006/")
                 .client(client)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build()
                 .create(ChatApiService.class);
-
         repo = new ChatRepository(api);
 
-        // load “all” → show one item per box
-        repo.loadAllMessages(/* yourUserId */16, new ChatRepository.CallbackFn<List<ChatMessageDto>>() {
+        // Load all messages, then show only the last per box
+        repo.loadAllMessages(16, new ChatRepository.CallbackFn<List<ChatMessageDto>>() {
             @Override public void onSuccess(List<ChatMessageDto> msgs) {
-                // group by boxId → pick last per box
-                List<ChatMessageDto> lastPerBox = ChatListAdapter.extractLastPerBox(msgs);
-                runOnUiThread(() -> adapter.updateData(lastPerBox));
+                List<ChatMessageDto> lastPerBox =
+                        ChatListAdapter.extractLastPerBox(msgs);
+                runOnUiThread(() ->
+                        adapter.updateData(lastPerBox)
+                );
             }
-            @Override public void onError(Throwable t) { /* show error */ }
+            @Override public void onError(Throwable t) {
+                // TODO: show an error toast/snackbar
+            }
         });
     }
 }
